working version without chat group messages

+++ app/crud.py
@@ -3,7 +3,7 @@ from uuid import uuid4
 from datetime import *
 from models import User, Message, Block, Group, GroupMember, ChatMetadata
 from schemas import UserCreate, MessageCreate, BlockCreate, GroupCreate, GroupMemberCreate, ChatMetadataCreate
-from boto3.dynamodb.conditions import Key
+from boto3.dynamodb.conditions import Key, Attr
 from config import config
 import boto3
 
@@ -34,12 +34,13 @@ def get_user_by_email(email):
 
 # Messages tables interfaces
 def create_message(message):
+    message['timestamp'] = message['timestamp'].isoformat()
     response = messages_table.put_item(Item=message)
     return response
 
 def get_user_chats(user_id):
     response = messages_table.query(
-        IndexName='user_id-index',
+        IndexName='sender_id-index',
         KeyConditionExpression=Key('sender_id').eq(user_id)
     )
     return response['Items']
@@ -47,7 +48,7 @@ def get_user_chats(user_id):
 def delete_old_messages(retention_hours):
     response = messages_table.scan()
     items = response['Items']
-    retention_time = datetime.utcnow() - timedelta(hours=retention_hours)
+    retention_time = datetime.now(timezone.utc) - timedelta(hours=retention_hours)
     old_messages = [item for item in items if int(item['timestamp']) < retention_time.timestamp()]
 
     for message in old_messages:
@@ -55,6 +56,7 @@ def delete_old_messages(retention_hours):
 
 # Chat Metadata table interfaces
 def create_chat_metadata(metadata):
+    metadata['latest_timestamp'] = metadata['latest_timestamp'].isoformat()
     response = chat_metadata_table.put_item(Item=metadata)
     return response
 
@@ -63,12 +65,12 @@ def get_chat_metadata(chat_id):
     return response.get('Item')
 
 def update_chat_metadata(chat_id, metadata):
+    metadata['latest_timestamp'] = metadata['latest_timestamp'].isoformat()
     response = chat_metadata_table.update_item(
         Key={'chat_id': chat_id},
-        UpdateExpression="set message_count=:m, start_index=:s, end_index=:e, latest_timestamp=:l",
+        UpdateExpression="set message_count=:m, end_index=:e, latest_timestamp=:l",
         ExpressionAttributeValues={
             ':m': metadata['message_count'],
-            ':s': metadata['start_index'],
             ':e': metadata['end_index'],
             ':l': metadata['latest_timestamp']
         },
@@ -85,6 +87,12 @@ def get_block(block_id):
     response = blocks_table.get_item(Key={'block_id': block_id})
     return response.get('Item')
 
+def get_block(sender_id, receiver_id):
+    response = blocks_table.scan(
+        FilterExpression=Attr('blocked_id').eq(sender_id) & Attr('blocker_id').eq(receiver_id)
+    )
+    return response.get('Item')
+
 # Group table interfaces
 def create_group(group):
     response = groups_table.put_item(Item=group)
@@ -95,17 +103,23 @@ def get_group(group_id):
     return response.get('Item')
 
 def add_user_to_group(group_member):
-    response = group_members_table.put_item(Item=group_member)
-    return response
-
+    if (get_group(group_member['group_id']) and get_user(group_member['user_id'])):
+        response = group_members_table.put_item(Item=group_member)
+        return response
+    else:
+        return None
+        
 def remove_user_from_group(group_id, user_id):
-    response = group_members_table.delete_item(
-        Key={
-            'group_id': group_id,
-            'user_id': user_id
-        }
-    )
-    return response
+    if (get_group(group_id)):
+        response = group_members_table.delete_item(
+            Key={
+                'group_id': group_id,
+                'user_id': user_id
+            }
+        )
+        return response
+    else:
+        return None
 
 def get_group_members(group_id):
     response = group_members_table.query(
